sjuste swagger e validations + Ajustes DTO e Exceptions

+++ users/src/main/java/com/utfpr/users/service/UserService.java
@@ -3,10 +3,12 @@ package com.utfpr.users.service;
 import com.utfpr.users.dto.UserRequestDTO;
 import com.utfpr.users.dto.UserResponseDTO;
 import com.utfpr.users.entity.User;
+import com.utfpr.users.exception.UserNotFoundException;
 import com.utfpr.users.repository.UserRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -24,10 +26,9 @@ public class UserService {
                 .collect(Collectors.toList());
     }
 
-    public UserResponseDTO getUserById(Long id) {
-        User user = userRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("User not found"));
-        return new UserResponseDTO(user.getId(), user.getUsername(), user.getEmail());
+    public Optional<UserResponseDTO> getUserById(Long id) {
+        return userRepository.findById(id)
+                .map(user -> new UserResponseDTO(user.getId(), user.getUsername(), user.getEmail()));
     }
 
     public UserResponseDTO createUser(UserRequestDTO userRequestDTO) {
@@ -41,7 +42,7 @@ public class UserService {
 
     public UserResponseDTO updateUser(Long id, UserRequestDTO userRequestDTO) {
         User user = userRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("User not found"));
+                .orElseThrow(() -> new UserNotFoundException("User with ID " + id + " not found."));
         user.setUsername(userRequestDTO.getUsername());
         user.setEmail(userRequestDTO.getEmail());
         User updatedUser = userRepository.save(user);
@@ -49,6 +50,9 @@ public class UserService {
     }
 
     public void deleteUser(Long id) {
+        if (!userRepository.existsById(id)) {
+            throw new UserNotFoundException("User with ID " + id + " not found.");
+        }
         userRepository.deleteById(id);
     }
 }
